Add geo.region to base/data results if queried, even if it's not existing
The response from the reader does not include geo.region in some cases (when geo is a region). As a hotfix for this problem, we are considering geo.region = geo if the client requested it and the response doesn't include it.

Also, it is guaranteed that every requested field will be present in the response, with a value or null