Add controls to the side on 'large' profile #394
Added controls to the side of 'large' profile.
When property button_expand is set to true all dialogs will show, when it is not set all will be hidden. When the button_expand property is set as an array it will display only those dialogs that are part of that array