Pop by age correct rendering and fixed extended grouping
new setting: group_offset. Sets the offset of the grouping relative to 0
e.g. group_by: 10, group_offset: 5 gives groups 5-14, 15-24, 25-34 etc (number of groups depends on chart domain and available data)

Bars are now correctly rendered, without overlap or gaps and correctly placed on the y-axis.
The domain is normally inclusive ([0,80] includes 80) but for this chart the end of the domain should be exclusive (i.e. [0,80) in interval notation). This is because the value 80 stands for anyone in their 81st life year. Could be changed of course to inclusive.

closes #516
touches on #273 (already closed)
closes #452
has a hack that needs to be extended or fixed for #392
closes #284 (or probably was already fixed before)
closes #286