Issue #751: Fix min/max axes
Fixes zooming when manually setting the min and max of axes.

Previously, when zooming with manually setting the axes, the range bump was not
taken into account. As a result, the min and max values were usually offset when
zooming with them. Essentially, zoomOnMaxMin used the fake min/max values, which
do not reflect the range bump.

Now, zoomOnMaxMin correctly takes into account the need to zoom using the real
rectangular coordinates that include the range bump.

Additionally, the fake axes values are correctly updated to account for the
range bump.