Issue #1169: Fix rectangle zoom ratio reset
When the difference between x1 and x2 is the same as the container's width, or
when the difference between y1 and y2 is the same as the container's height,
_zoomOnRectangle will set zoom to 1, but change the value of ratioX or ratioY
when trying to zoom.

However, in the zoom function, a hack is in place to reset the ratios to 1 when
zoom is 1, which prevents zooming under the above conditions.

If the above hack is removed, then it becomes possible to zoom in under the
above conditions, but impossible to zoom out with the mousewheel because zoom
is already at 1.

This commit changes the ratio reset conditions such that the ratios are only
reset when zooming out with the mousewheel. Doing so enables zooming in under
the above conditions and zooming out.