Added ObjectLeafs property and isObjectLeaf(name) function to Model
Now it's possible to have ModelLeafs with values that are Javascript Objects. This enables for example filters, written in JSON, which become Objects when parsed, to not be seen as part of the Model. Before, any value that was an Object was automatically another branch in the model.
So entities.show is from now on a ModelLeaf with a value which is a javascript object. This means that when you change it, it will fire off a change event.

This fixes the minimap because changing entities.show before didn't trigger a change event (because adding new submodels (i.e. changing the Model-ified filter) in runtime is unexpected and doesn't trigger a change event).