Remove prototypal inheritance triggering case
When you define a class, you can add variables with value to the paremeter of the `extend` function.
Normally, you'd expect this to be the instantiated value when creating an Object of that class. But in our JS: The variable becomes part of the prototype of the the newly created class.
When editing the variable somewhere in the code then, you're actually changing the prototype
When the class gets extended to some other class, it will inherit the changed prototype, including possibly new values assigned to it in the code (unless it is overloaded in superclass).

So this is NOT the same as normal OOP. There a change later in the code wouldn't affect extended classes. There is no shared prototype.

e.g.
I define `model_bind` on Component in `extend`: https://github.com/vizabi/vizabi/blob/f7b0b98681efe09e194a2afc48ad679c1fc7a6e7/src/base/component.js#L14
Further down I set `model_bind` to something in the code:
https://github.com/vizabi/vizabi/blob/f7b0b98681efe09e194a2afc48ad679c1fc7a6e7/src/base/component.js#L49
This happens for example when Tool is created, `model_binds` will be filled with binds from external page.
`this.model_binds` will contain external page binds. But since this.model_binds refers to `Component`'s prototype, the prototype contains the same `model_binds`!

When sometime later, you extend component, e.g. for `moreoptions` dialog, which doesn't have it's own `this.model_binds` declaration, `this.model_binds` of `moreoptions` will point to its prototype, which is component's model_binds. Which in this case was all the external page binds.
So then a lot of componenets all of a sudden have external page bindings!

In classic OOP, the class declaration would just be a blueprint for the object. Once the object is created you can change its variables without it influencing variables in children, other objects etc. So that's not the case here.
In this case its is kind of like a static variable which you can approach through `this`. Yikes!