Consider lastModified time when caching the data
Caching data in CSV reader and in DATASTORAGE can lead to problems when file was modified and opened again without restarting vizabi. This happens often in Gapminder offline app when users would open the same file with some modifications many times until they are happy with the file format. Usually we are caching files and queries to readers only considering the data path, but we ignore the fact if a file was modified. This commit adds lastModified property to caching instructions both on CSV reader level and on DATASTORAGE level. Reader itself is not able to identify modified date of a file, but if external page or app is able to they can use that new property of “data” config